edit: slide부분 setInterval구현중
