+++ src/pages/flowcard/card.scss
@@ -1,0 +1,13 @@
+.flowcard-wrapper {
+  box-sizing: border-box;
+  .flowcard-img {
+    width: 230px;
+    height: 230px;
+    border-radius: 70px 70px 70px 70px;
+    background-color: blue;
+    margin-right: 30px;
+    img {
+      margin: 0 auto;
+    }
+  }
+}
