--- src/pages/flowcard/folwcard.js
@@ -1,88 +1,0 @@
-import './flowcard.scss';
-
-export default function FlowCard() {
-  return (
-    <div className="flowcard-wrapper">
-      <section className="product-card-container">
-        <div className="product-card">
-          <div className="product-flowcard-toplist">
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide1.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee2.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee3.jpg" alt="coffee3" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee1.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee2.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee3.jpg" alt="coffee3" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee1.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee2.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee3.jpg" alt="coffee3" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee1.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee2.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee3.jpg" alt="coffee3" />
-            </div>
-          </div>
-          <div className="product-flowcard-bottomlist">
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide1.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide2.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide3.jpg" alt="coffee3" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide4.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./cimages/slide/slide5.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide6.jpg" alt="coffee3" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide7.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide1.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide2.jpg" alt="coffee3" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./images/slide/slide3.jpg" alt="coffee1" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee2.jpg" alt="coffee2" />
-            </div>
-            <div className="product-flowcard-item">
-              <img src="./coffee3.jpg" alt="coffee3" />
-            </div>
-          </div>
-        </div>
-      </section>
-    </div>
-  );
-}
